Restructure pipeline and avoid circular imports
