+++ asgardpy/__init__.py
@@ -1,4 +1,12 @@
-from . import analysis, config, data, io
+from . import analysis, base, config, data, io
 from .version import VERSION, VERSION_SHORT
 
-__all__ = ["analysis", "data", "io", "config", "VERSION", "VERSION_SHORT"]
+__all__ = [
+    "analysis",
+    "base",
+    "data",
+    "io",
+    "config",
+    "VERSION",
+    "VERSION_SHORT",
+]
