+++ asgardpy/config/__init__.py
@@ -1,3 +1,3 @@
-from asgardpy.config.generator import AsgardpyConfig, recursive_merge_dicts
+from asgardpy.config.generator import AsgardpyConfig
 
-__all__ = ["AsgardpyConfig", "recursive_merge_dicts"]
+__all__ = ["AsgardpyConfig"]
