+++ docs/source/base_base.rst
@@ -1,0 +1,7 @@
+asgardpy.base.base module
+=========================
+
+.. automodule:: asgardpy.base.base
+   :members: BaseConfig, AnalysisStep, AnalysisStepBase
+   :undoc-members:
+   :show-inheritance:
