--- docs/source/data_base.rst
@@ -1,7 +1,0 @@
-asgardpy.data.base module
-=========================
-
-.. automodule:: asgardpy.data.base
-   :members: BaseConfig, AnalysisStep
-   :undoc-members:
-   :show-inheritance:
